[MMU3] Fix nozzle double parking.
In very rare conditions, i.e. a second error arise while trying to fix the first one, even though thee nozzle is already parked, it lifts in the Z axis a second time, causing the nozzle not to be able to return to its original Z height when the errors are solved (there will be an offset of +NOZZLE_PARK_POINT.z). This commit fixes that.