[MMU3] Fix for print getting cancelled after `MAX_RETRIES` is reached.
Fixed `MMU2::ToolChangeCommonOnce()` to prevent the print getting cancelled after the full screen error message is displayed `MAX_RETRIES` amount. With this modification, the full screen error message is shown on the `MAX_RETRIES` retry as usual, but then keeps showing the full screen message until the problem is solved, where as in the previous implementation the print is cancelled without any notification.